requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
python-dotenv>=1.0.0
deepgram-sdk>=3.7.0
elevenlabs>=1.9.0
tavily-python>=0.5.0
mem0ai>=0.1.29
requests>=2.32.0
pydantic>=2.10.0
httpx>=0.27.0
google-genai>=1.0.0
groq>=0.9.0